feat: Atualizacao no front das pages perfil,historico quedas
