+++ package.json
@@ -9,22 +9,24 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~3.2.1",
+    "@expo/metro-runtime": "~3.2.3",
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.10.1",
-    "expo": "~51.0.21",
+    "@react-navigation/stack": "^6.4.1",
+    "@types/react": "~18.2.79",
+    "expo": "~51.0.28",
     "expo-status-bar": "~1.12.1",
     "firebase": "^10.12.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.74.3",
+    "react-native": "^0.74.5",
     "react-native-paper": "^5.12.3",
+    "react-native-reanimated": "^3.15.0",
     "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
     "react-native-web": "~0.19.10",
-    "typescript": "~5.3.3",
-    "@types/react": "~18.2.79"
+    "typescript": "~5.3.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
