--- tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "compilerOptions": {},
-  "extends": "expo/tsconfig.base"
-}
